requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pytest==7.4.0
pytest-mock==3.11.1
flake8==6.0.0
